// src/generated/fontManifest.ts
export type FontOption = {
  id: string;
  label: string;
  family: string;
  file: string;
  sourceFile: string;
};

const fontAsset0 = new URL("../../manchufonts/001A-Buleku-R.ttf", import.meta.url).href;
const fontAsset1 = new URL("../../manchufonts/002A-Buleku-VT.ttf", import.meta.url).href;
const fontAsset2 = new URL("../../manchufonts/002A-Buleku-XX.ttf", import.meta.url).href;
const fontAsset3 = new URL("../../manchufonts/002A-M-Trans-R-VU.ttf", import.meta.url).href;
const fontAsset4 = new URL("../../manchufonts/Sungar Ginggulere hergen.ttf", import.meta.url).href;
const fontAsset5 = new URL("../../manchufonts/Sungar PaKa.ttf", import.meta.url).href;
const fontAsset6 = new URL("../../manchufonts/Sunggar Kara jijungga.ttf", import.meta.url).href;
const fontAsset7 = new URL("../../manchufonts/Sunggar Xayan jijungga - Bold.ttf", import.meta.url).href;
const fontAsset8 = new URL("../../manchufonts/Sunggar Xayan jijungga.ttf", import.meta.url).href;
const fontAsset9 = new URL("../../manchufonts/Sunggar Xayan jijungga(joringga) - Bold.ttf", import.meta.url).href;
const fontAsset10 = new URL("../../manchufonts/Sunggar Xayan jijungga(joringga).ttf", import.meta.url).href;
const fontAsset11 = new URL("../../manchufonts/XM_BiaoHei.ttf", import.meta.url).href;
const fontAsset12 = new URL("../../manchufonts/XM_GuFeng.ttf", import.meta.url).href;
const fontAsset13 = new URL("../../manchufonts/XM_LiuYe.ttf", import.meta.url).href;
const fontAsset14 = new URL("../../manchufonts/XM_ShuKai.ttf", import.meta.url).href;
const fontAsset15 = new URL("../../manchufonts/XM_WenJian.ttf", import.meta.url).href;
const fontAsset16 = new URL("../../manchufonts/XM_WenQin.ttf", import.meta.url).href;
const fontAsset17 = new URL("../../manchufonts/XM_XingShu.ttf", import.meta.url).href;
const fontAsset18 = new URL("../../manchufonts/XM_YaBai.ttf", import.meta.url).href;
const fontAsset19 = new URL("../../manchufonts/XM_YingBi.ttf", import.meta.url).href;
const fontAsset20 = new URL("../../manchufonts/XM_YuJinZhuan.ttf", import.meta.url).href;
const fontAsset21 = new URL("../../manchufonts/XM_ZhengBai.ttf", import.meta.url).href;
const fontAsset22 = new URL("../../manchufonts/XM_ZhengHei.ttf", import.meta.url).href;

export const FONT_OPTIONS: FontOption[] = [
  {
    id: "001a-buleku-r",
    label: "001A Buleku R",
    family: "ManchuPreview01",
    file: fontAsset0,
    sourceFile: "001A-Buleku-R.ttf",
  },
  {
    id: "002a-buleku-vt",
    label: "002A Buleku VT",
    family: "ManchuPreview02",
    file: fontAsset1,
    sourceFile: "002A-Buleku-VT.ttf",
  },
  {
    id: "002a-buleku-xx",
    label: "002A Buleku XX",
    family: "ManchuPreview03",
    file: fontAsset2,
    sourceFile: "002A-Buleku-XX.ttf",
  },
  {
    id: "002a-m-trans-r-vu",
    label: "002A M Trans R VU",
    family: "ManchuPreview04",
    file: fontAsset3,
    sourceFile: "002A-M-Trans-R-VU.ttf",
  },
  {
    id: "sungar-ginggulere-hergen",
    label: "Sungar Ginggulere hergen",
    family: "ManchuPreview05",
    file: fontAsset4,
    sourceFile: "Sungar Ginggulere hergen.ttf",
  },
  {
    id: "sungar-paka",
    label: "Sungar PaKa",
    family: "ManchuPreview06",
    file: fontAsset5,
    sourceFile: "Sungar PaKa.ttf",
  },
  {
    id: "sunggar-kara-jijungga",
    label: "Sunggar Kara jijungga",
    family: "ManchuPreview07",
    file: fontAsset6,
    sourceFile: "Sunggar Kara jijungga.ttf",
  },
  {
    id: "sunggar-xayan-jijungga-bold",
    label: "Sunggar Xayan jijungga Bold",
    family: "ManchuPreview08",
    file: fontAsset7,
    sourceFile: "Sunggar Xayan jijungga - Bold.ttf",
  },
  {
    id: "sunggar-xayan-jijungga",
    label: "Sunggar Xayan jijungga",
    family: "ManchuPreview09",
    file: fontAsset8,
    sourceFile: "Sunggar Xayan jijungga.ttf",
  },
  {
    id: "sunggar-xayan-jijungga-joringga-bold",
    label: "Sunggar Xayan jijungga(joringga) Bold",
    family: "ManchuPreview10",
    file: fontAsset9,
    sourceFile: "Sunggar Xayan jijungga(joringga) - Bold.ttf",
  },
  {
    id: "sunggar-xayan-jijungga-joringga",
    label: "Sunggar Xayan jijungga(joringga)",
    family: "ManchuPreview11",
    file: fontAsset10,
    sourceFile: "Sunggar Xayan jijungga(joringga).ttf",
  },
  {
    id: "xm-biaohei",
    label: "XM BiaoHei",
    family: "ManchuPreview12",
    file: fontAsset11,
    sourceFile: "XM_BiaoHei.ttf",
  },
  {
    id: "xm-gufeng",
    label: "XM GuFeng",
    family: "ManchuPreview13",
    file: fontAsset12,
    sourceFile: "XM_GuFeng.ttf",
  },
  {
    id: "xm-liuye",
    label: "XM LiuYe",
    family: "ManchuPreview14",
    file: fontAsset13,
    sourceFile: "XM_LiuYe.ttf",
  },
  {
    id: "xm-shukai",
    label: "XM ShuKai",
    family: "ManchuPreview15",
    file: fontAsset14,
    sourceFile: "XM_ShuKai.ttf",
  },
  {
    id: "xm-wenjian",
    label: "XM WenJian",
    family: "ManchuPreview16",
    file: fontAsset15,
    sourceFile: "XM_WenJian.ttf",
  },
  {
    id: "xm-wenqin",
    label: "XM WenQin",
    family: "ManchuPreview17",
    file: fontAsset16,
    sourceFile: "XM_WenQin.ttf",
  },
  {
    id: "xm-xingshu",
    label: "XM XingShu",
    family: "ManchuPreview18",
    file: fontAsset17,
    sourceFile: "XM_XingShu.ttf",
  },
  {
    id: "xm-yabai",
    label: "XM YaBai",
    family: "ManchuPreview19",
    file: fontAsset18,
    sourceFile: "XM_YaBai.ttf",
  },
  {
    id: "xm-yingbi",
    label: "XM YingBi",
    family: "ManchuPreview20",
    file: fontAsset19,
    sourceFile: "XM_YingBi.ttf",
  },
  {
    id: "xm-yujinzhuan",
    label: "XM YuJinZhuan",
    family: "ManchuPreview21",
    file: fontAsset20,
    sourceFile: "XM_YuJinZhuan.ttf",
  },
  {
    id: "xm-zhengbai",
    label: "XM ZhengBai",
    family: "ManchuPreview22",
    file: fontAsset21,
    sourceFile: "XM_ZhengBai.ttf",
  },
  {
    id: "xm-zhenghei",
    label: "XM ZhengHei",
    family: "ManchuPreview23",
    file: fontAsset22,
    sourceFile: "XM_ZhengHei.ttf",
  }
];